# Optimized Render code for TextBuffers