--- a/src/main/scala/li/cil/oc/client/renderer/TextBufferRenderCache.scala
+++ b/src/main/scala/li/cil/oc/client/renderer/TextBufferRenderCache.scala
@@ -16,9 +16,8 @@ import net.minecraft.tileentity.TileEntity
 import org.lwjgl.opengl.GL11
 
 object TextBufferRenderCache extends Callable[Int] with RemovalListener[TileEntity, Int] {
-  val renderer =
-    if (Settings.get.fontRenderer == "texture") new font.StaticFontRenderer()
-    else new font.DynamicFontRenderer()
+	
+  final val renderer = new font.TextureFontRenderer()
 
   private val cache = com.google.common.cache.CacheBuilder.newBuilder().
     expireAfterAccess(2, TimeUnit.SECONDS).
@@ -26,54 +25,13 @@ object TextBufferRenderCache extends Callable[Int] with RemovalListener[TileEnti
     asInstanceOf[CacheBuilder[TextBufferRenderData, Int]].
     build[TextBufferRenderData, Int]()
 
-  // To allow access in cache entry init.
-  private var currentBuffer: TextBufferRenderData = _
-
   // ----------------------------------------------------------------------- //
   // Rendering
   // ----------------------------------------------------------------------- //
 
   def render(buffer: TextBufferRenderData) {
-    currentBuffer = buffer
-    compileOrDraw(cache.get(currentBuffer, this))
-  }
-
-  private def compileOrDraw(list: Int) = {
-    if (currentBuffer.dirty) {
-      RenderState.checkError(getClass.getName + ".compileOrDraw: entering (aka: wasntme)")
-
-      for (line <- currentBuffer.data.buffer) {
-        renderer.generateChars(line)
-      }
-
-      val doCompile = !RenderState.compilingDisplayList
-      if (doCompile) {
-        currentBuffer.dirty = false
-        GL11.glNewList(list, GL11.GL_COMPILE_AND_EXECUTE)
-
-        RenderState.checkError(getClass.getName + ".compileOrDraw: glNewList")
-      }
-
-      renderer.drawBuffer(currentBuffer.data, currentBuffer.viewport._1, currentBuffer.viewport._2)
-
-      RenderState.checkError(getClass.getName + ".compileOrDraw: drawString")
-
-      if (doCompile) {
-        GL11.glEndList()
-
-        RenderState.checkError(getClass.getName + ".compileOrDraw: glEndList")
-
-      }
-
-      RenderState.checkError(getClass.getName + ".compileOrDraw: leaving")
-
-      true
-    }
-    else {
-      GL11.glCallList(list)
-
-      RenderState.checkError(getClass.getName + ".compileOrDraw: glCallList")
-    }
+    var reRender = cache.getIfPresent(buffer) == null
+    renderer.drawBuffer(buffer.data, buffer.viewport._1, buffer.viewport._2, cache.get(buffer, this), reRender)
   }
 
   // ----------------------------------------------------------------------- //
@@ -82,19 +40,19 @@ object TextBufferRenderCache extends Callable[Int] with RemovalListener[TileEnti
 
   def call = {
     RenderState.checkError(getClass.getName + ".call: entering (aka: wasntme)")
-
-    val list = GLAllocation.generateDisplayLists(1)
-    currentBuffer.dirty = true // Force compilation.
+    
+    var texID = font.TextureFontRenderer.createTexture
 
     RenderState.checkError(getClass.getName + ".call: leaving")
-
-    list
+    
+    texID
   }
+
 
   def onRemoval(e: RemovalNotification[TileEntity, Int]) {
     RenderState.checkError(getClass.getName + ".onRemoval: entering (aka: wasntme)")
 
-    GLAllocation.deleteDisplayLists(e.getValue)
+    GL11.glDeleteTextures(e.getValue)
 
     RenderState.checkError(getClass.getName + ".onRemoval: leaving")
   }