--- a/src/main/scala/li/cil/oc/client/renderer/markdown/segment/CodeSegment.scala
+++ b/src/main/scala/li/cil/oc/client/renderer/markdown/segment/CodeSegment.scala
@@ -7,8 +7,6 @@ import org.lwjgl.opengl.GL11
 
 private[markdown] class CodeSegment(val parent: Segment, val text: String) extends BasicTextSegment {
   override def render(x: Int, y: Int, indent: Int, maxWidth: Int, renderer: FontRenderer, mouseX: Int, mouseY: Int): Option[InteractiveSegment] = {
-    TextBufferRenderCache.renderer.generateChars(text.toCharArray)
-
     var currentX = x + indent
     var currentY = y
     var chars = text
@@ -17,7 +15,7 @@ private[markdown] class CodeSegment(val parent: Segment, val text: String) exten
     while (chars.length > 0) {
       val part = chars.take(numChars)
       GL11.glColor4f(0.75f, 0.8f, 1, 1)
-      TextBufferRenderCache.renderer.drawString(part, currentX, currentY)
+      // TODO TextBufferRenderCache.renderer.drawString(part, currentX, currentY)
       currentX = x + wrapIndent
       currentY += lineHeight(renderer)
       chars = chars.drop(numChars).dropWhile(_.isWhitespace)