… src/main/scala/li/cil/oc/client/renderer/font/DynamicFontRenderer.scala
… src/main/scala/li/cil/oc/client/renderer/font/StaticFontRenderer.scala