Fixes #779

This PR replaces the existing infrastructure for TextBuffer Rendering with a texture sheet.

Benefits:
* Significantly less OpenGL calls
* Dramatic increase in FPS on certain GPUS (see below for stats)
* Only re-draws the parts of the screen that need it
* Viewing the screen both in world and in window draw a single quad each

TODO:
* Changing screen resolution does not stretch as expected.  I would argue that this is a feature, but it could be considered a bug

Stress Test Program (Courtesy of Ristelle):
```
local component = require("component")
local gpu = component.gpu
local r = {"00","33","66","99","cc","ff"}
local g = {"00","24","49","6d","92","b6","db","ff"}
local b = {"00","40","80","c0","ff"}
local x,y = gpu.getResolution()
local rr = 1
local rg = 1
local rb = 1
while true do
    rr = rr+ 1
    if rr == 6 then
        rr = 1
        rg = rg + 1
    end
    if rg == 8 then
        rg = 1
        rb = rb +1
    end
    if rb == 5 then
        rb = 1
    end
    local rgb = string.format("0x%s%s%s",r[rr],g[rg],b[rb])
    gpu.setBackground(tonumber(rgb))
    print("Hello Cam asjdkfasdhjkhfakh Ｆｏｒ ｅｘａｍｐｌｅ， ｔｈｉｓ．12839054812908902!")
    os.sleep(0)
end
```

Test Results:
AMD R9 380: 5FPS - 60FPS (linux vsync locked)
NVIDIA GTX 650: 80FPS - 80FPS
more coming soon