@Ristelle 2018-06-18
I'll like to mentions some more details as well.  
Some users may expreince poorer performances  
Cam Test:  
![2018-06-18_21 11 11](https://user-images.githubusercontent.com/7894419/41538754-9162b83a-733e-11e8-9421-17de301ab127.png)  
Main Branch:
![2018-06-18_21 23 38](https://user-images.githubusercontent.com/7894419/41538776-9e0eeea0-733e-11e8-9648-ae01b4466f89.png)