@Ristelle 2018-06-18
Oh wow I can upload the Test world... Sure!  
[Testworld.zip](https://github.com/MightyPirates/OpenComputers/files/2112044/Testworld.zip)