@payonel 2018-09-09
if we've added this to the our 1.7.3 milestone, we need to have the PR author finish the concerns of the review.

@cam72cam before we can merge this, we would need all of the following addressed

1. how GPU memory usage reads comparing 32 screen with superflat scenes? GPU-Z should probably work.

2. keep the old method avaialble via config until it's proven

3. Ristelle demonstrated on nvidia 650 we drop from 91fps to 52fps? That seems bad, doesn't it?

4. asie would like to see making linear scaling the min filter for screens.

5. asie is concerned about vram usage at scale. @asiekierka: what scale value matters?