@cam72cam 2018-09-09
1. I am on linux and don't have GPU-Z
2. I'll try to see if I can figure out a way to do this
3. That's a fair comparison, a lot of the performance of this code depends on the GPU you are using.  Intel and AMD seem to work better with my PR.  NVIDIA is a toss up IIRC
4: I'll see what I can do
5: aise?