@Vexatos 2018-09-09
For what it's worth, for NVIDIA GPUs on Linux (with the official drivers), a very useful tool is running `nvidia-smi -l` to see basically all relevant data at a glance.