@payonel 2018-11-10
@cam72cam I can confirm there is a drop in FPS to half on integrated graphics. I think this will affect a good number of users, unfortunately
I think this speedup, however, would be a welcome improvement. I think the right thing going forward is to provide a configuration option to enable this.

We are very close to releasing our 1.7.3 patch. I choice on this PR is three-fold
1. I'm removing the 1.7.3 miletone from this PR
2. I think this change in behavior should be configurable, but off by default. We should continue using the old code by default.
3. I think it is okay to only upgrade our 1.12 branch with this feature. I'm not interested in making these improvements to the lower versions